I really tried SO hard.
I ran everything and the brute force tables are truly not going to happen for me. I had a rough time trying to understand that portion. HTML works (not the cutest). I know this isn't what most would consider to be great but I learned alot from struggling through this and hopefully I can figure this out as I will continue to work on it. Please contact me if it isnt running properly, my code sometimes has issues on pycharm and I run it from the terminal. Hope this is somewhat decent. I had alot of issues with iterating as it kept breaking my code, hope to have your help on that. Thank you.